+++ src/app/window_manager.rs
@@ -1350,6 +1350,7 @@ impl WindowManager {
                     updated |= renderer.update_font_antialias(config.font_antialias);
                     updated |= renderer.update_font_hinting(config.font_hinting);
                     updated |= renderer.update_font_thin_strokes(config.font_thin_strokes);
+                    updated |= renderer.update_minimum_contrast(config.minimum_contrast);
                     if updated {
                         window_state.needs_redraw = true;
                     }